Extract test-selection helper and make --level deselect
